[Refactor] ViewModel 적용하기
- 계산 로직을 ViewModel이 담당
